moved function to its own .ts file
