+++ source/win-main/file-manager/util/alphabeticSort.ts
@@ -1,0 +1,21 @@
+import { MDFileMeta, CodeFileMeta, DirMeta } from '@dts/common/fsal'
+
+export default function alphabeticSort(dirArray: any, sortBy: string): DirMeta[] {
+    function compare(a: any, b: any) {
+
+      let nameA = a.name.toLowerCase(), nameB = b.name.toLowerCase();
+
+      // if workspaces have the same name, compare their parent's name
+      if (nameA == nameB) {
+        let dirA = a.dir.toLowerCase(), dirB = b.dir.toLowerCase();
+        if (dirA < dirB)
+          return -1;
+        if (dirA > dirB)
+          return 1;
+      } 
+      return 0;
+    }
+    //let temp = this.getFilteredTree.filter(item => item.type === 'directory');
+    if (sortBy == "desc") return dirArray.sort(compare).reverse() as DirMeta[];
+    return dirArray.sort(compare) as DirMeta[];
+  }
